apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.19
requests==2.32.3
python-dotenv==1.2.1
